refactor: 탐색 관련 폴더 정리 및 전시목록 카드 디자인 수정
