+++ src/components/search/index.ts
@@ -1,6 +1,5 @@
-export { ExhibitionCard } from './ExhibitionCard';
-export { FilterChip } from './FilterChip';
-export { FilterModal } from './FilterModal';
+export { FilterChip } from './filter/FilterChip';
+export { FilterModal } from './filter/FilterModal';
 export {
   DEFAULT_FILTER_STATE,
   FIELD_OPTIONS,
@@ -11,6 +10,10 @@ export {
   REGION_OPTIONS,
   STATUS_OPTIONS,
   TYPE_OPTIONS,
-} from './filterOptions';
+} from './filter/filterOptions';
+export { ExhibitionCard } from './list/ExhibitionCard';
+export { ExhibitionMap } from './map/ExhibitionMap';
+export { ExhibitionMapCard } from './map/ExhibitionMapCard';
+export { ExhibitionMapMarker } from './map/ExhibitionMapMarker';
 export type { Exhibition } from './types';
 export type { FilterState, FilterTab } from '@/types/search';
